Source code synced with Eugene game development kit 8.7.8

+++ source/graygdk.cpp
@@ -2157,7 +2157,7 @@ namespace GRAYGDK
    return this->is_software_render();
   }
 
-  unsigned long int Screen::get_color() const
+  unsigned int Screen::get_color() const
   {
    return this->get_depth();
   }
